feat: change behavior of "refresh" button
The "refresh" button has been wrongly interpreted both by the
certificate holders (users) and the people checking the certificates
(scanners).

Nowadays when visiting a venue that requires a COVID Certificate
verification the certificate will not be scanned with the official
COVID Certificate Check app, but rather the scanner will press the
"refresh" button on the user phone.

This obviously defeats the purpose of having a cryptographically secure
QR code signed by a trusted party.

This behavior has been adopted due to a mistake in the UI / UX design,
where clicking the "refresh" button makes the QR code and the
information bubbles go green for a couple of seconds, saying
"verification successful".

A misinterpretation of the feature has thus caused an extremely
wrong and dangerous pattern when scanning the certificates.

This commit partially fixes the issue by adding a notice
_every time_ that button is pressed, and by hiding the green
verification icons / colors. An invalid certificate will anyways
result in a red background as usual.